backend/requirements.txt
fastapi[standard]
SQLAlchemy
pydantic
python-multipart

# Auth
python-jose[cryptography]
passlib[bcrypt]

# Env support
python-dotenv